+++ JotunnLib/Main.cs
@@ -20,7 +20,7 @@ namespace Jotunn
         /// <summary>
         ///     The current version of the Jotunn library.
         /// </summary>
-        public const string Version = "2.4.8";
+        public const string Version = "2.4.9";
 
         /// <summary>
         ///     The name of the library.